폴더 및 namespace 정리
