+++ TLKAPP/Views/FilterControl.xaml
@@ -59,7 +59,6 @@
             <TextBox Grid.Column="1" Text="{Binding View.ReplaceText, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
                      FontSize="16" FontFamily="{Binding Settings.FontFamilyName}"/>
         </Grid>
-
     </Grid>
 
 </UserControl>
