+++ TLKAPP/Views/MenuControl.xaml
@@ -1,20 +1,18 @@
 ﻿<UserControl x:Class="TLKAPP.Views.MenuControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:app="clr-namespace:TLKAPP"
              xmlns:vc="clr-namespace:TLKAPP.ViewConverter"
              xmlns:tlkcmd="clr-namespace:TLKVIEWMODLES.Commands;assembly=TLKVIEWMODLES"
-             app:OpenFileDlg.Show="{Binding OpenFileDlg}"
-             app:FontDlg.Show="{Binding FontDlg}"
-             app:MsgBox.Show="{Binding MsgBox}">
+             vc:OpenDlg.Show="{Binding OpenDlg}" 
+             vc:FontDlg.Show="{Binding FontDlg}"
+             vc:MsgBox.Show="{Binding MsgBox}">
     <Grid>
         <Menu>
             <MenuItem Header="파일">
                 <MenuItem Header="열기" Command="{tlkcmd:LoadFileCommand}"/>
                 <MenuItem Header="닫기" Command="{tlkcmd:UnloadFileCommand}"/>
                 <Separator/>
-                <MenuItem Header="종료" Command="{tlkcmd:WinCloseCommand}" 
-                          CommandParameter="{Binding Converter={vc:WinCloseArgs}}"/>
+                <MenuItem Header="종료" Command="{tlkcmd:WinCloseCommand}" CommandParameter="{Binding Converter={vc:WinCloseArgs}}"/>
             </MenuItem>
             <MenuItem Header="도구">
                 <MenuItem Header="Encoding">
