+++ TLKVIEWMODLES/TLKVIEWMODLES.csproj
@@ -59,15 +59,12 @@
   <ItemGroup />
   <ItemGroup>
     <Compile Include="Commands\BaseCommands.cs" />
+    <Compile Include="Commands\MenuCommands.cs" />
     <Compile Include="Commands\ViewCommands.cs" />
     <Compile Include="Contexts\BaseContext.cs" />
-    <Compile Include="Commands\MenuCommands.cs" />
-    <Compile Include="Contexts\Models\TLKENTRY.cs" />
-    <Compile Include="Contexts\Models\TLKTEXT.cs" />
     <Compile Include="Contexts\SettingsContext.cs" />
-    <Compile Include="Contexts\Models\TLKTextCollection.cs" />
     <Compile Include="Contexts\ViewContext.cs" />
-    <Compile Include="Contexts\Models\ViewTabModels.cs" />
+    <Compile Include="ViewModels\ViewTabModels.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Type\CommandArgsType.cs" />
     <Compile Include="Type\ViewTpye.cs" />
